+++ src/lib/ui/template.ts
@@ -1,168 +1,33 @@
+import * as config from 'lib/config'
+import * as datetime from 'lib/datetime'
+import * as fileHelpers from 'lib/fileHelpers'
+
 import * as express from 'express'
 import * as fs from 'fs'
-import * as config from 'lib/config'
-import * as dateTime from 'lib/datetime'
-import * as fileHelpers from 'lib/filehelpers'
+
+import * as log4js from 'log4js'
 import * as path from 'path'
 
-import * as vm from 'vm'
-
 /*tslint:disable*/
-const svelte = require('svelte')
+require('svelte/ssr/register')
+const {Store} = require('svelte/store.umd.js')
 /*tslint:enable*/
 
-interface AST {
-	children: AST[]
-	name: string
-	type: string
-}
-
-interface ParseError {
-	frame: string
-	loc: {
-		column: number
-		line: number
-	}
-	message: string
-	name: string
-}
-
-interface Renderer {
-	render(props?: object): {html: string}
-}
-
 const rootDir = process.cwd()
-const componentDir = path.join(rootDir, 'component')
+// const componentDir = path.join(rootDir, 'component')
 const pageDir = path.join(rootDir, 'page')
 
-let allComponentNames = ''
+// const componentList: Record<string, string> = {}
 
-let componentList: {[name: string]: string} = {}
+const logger = log4js.getLogger('svelte compilation')
 
-let components: {[key: string]: Renderer | undefined} = {}
-let currentRequest: express.Request
-let pages: {[key: string]: Renderer} = {}
-
-function compile(
-	path: string,
-	source: string,
-	name: string,
-	isComponent?: boolean
-): Renderer | undefined {
-	// Ensure that there aren't any <script> blocks with content.
-	svelte.preprocess(source, {
-		script: (info: {content: string}): {code: string} => {
-			if (info.content) {
-				throw new Error(`ui/svelte: unexpected <script> tag found in ${path}`)
-			}
-			return {code: ''}
-		},
-	})
-	let componentNames = allComponentNames
-	let constructorName = name
-	if (isComponent) {
-		const componentSet = new Set(Object.keys(componentList))
-		componentSet.delete(name)
-		componentNames = Array.from(componentSet).join(', ')
-	} else {
-		constructorName = name + 'Page'
-	}
-	source += `
-<script>
-export default {
-components: {${componentNames}},
-data() {
-    const req = getCurrentRequest()
-    return {
-        config: configModule,
-        currentReq: req,
-        datetime: dateTimeModule,
-        fileHelpers,
-        i18n: req.__ ? req.__.bind(req) : null,
-        signedInUser: req.user,
-        toHtml: toHtml,
-    }
-}
-}
-</script>
-`
-	let compiled
-	try {
-		compiled = svelte.compile(source, {
-			css: false,
-			dev: false,
-			filename: path,
-			format: 'cjs',
-			generate: 'ssr',
-			name: constructorName,
-			onwarn: () => {
-				return
-			},
-		})
-	} catch (err) {
-		logParseError(path, err)
-		return
-	}
-	return createModule(path, compiled.js.code, componentNames) as Renderer
-}
-
-function createModule(filename: string, code: string, componentNames: string) {
-	const module = {exports: {}}
-	const wrapper = vm.runInThisContext(
-		`(function(module, exports, require, components, getCurrentRequest, configModule, dateTimeModule,
-		 fileHelpers, toHtml) {
-const {${componentNames}} = components
-${code}
-});`,
-		{filename}
-	)
-	wrapper(
-		module,
-		module.exports,
-		require,
-		components,
-		getCurrentRequest,
-		config,
-		dateTime,
-		fileHelpers,
-		toHtml
-	)
-	return module.exports
-}
-
-function gatherComponents(node: AST, seen: Set<string>) {
-	if (node.type === 'Element' && isCapitalised(node.name)) {
-		seen.add(node.name)
-	}
-	if (node.children && node.children.length) {
-		for (const child of node.children) {
-			gatherComponents(child, seen)
-		}
-	}
-}
-
-function getCurrentRequest() {
-	return currentRequest
-}
-
-function getName(ident: string) {
-	ident = ident.split('/').pop()!.replace(/-/g, '')
-	return ident.charAt(0).toUpperCase() + ident.slice(1)
-}
-
-function getDependencies(root: AST) {
-	const seen: Set<string> = new Set()
-	gatherComponents(root, seen)
-	return seen
-}
-
-function isCapitalised(ident: string) {
-	if (!ident) {
-		return false
-	}
-	const char = ident.charCodeAt(0)
-	return char >= 65 && char <= 90
-}
+// function isDirectory(path: string) {
+// 	try {
+// 		return fs.statSync(path).isDirectory()
+// 	} catch (err) {
+// 		return false
+// 	}
+// }
 
 function isFile(path: string) {
 	try {
@@ -170,18 +35,6 @@ function isFile(path: string) {
 	} catch (err) {
 		return false
 	}
-}
-
-function isDirectory(path: string) {
-	try {
-		return fs.statSync(path).isDirectory()
-	} catch (err) {
-		return false
-	}
-}
-
-function componentIdentity(component: string): [string, string] {
-	return [component.slice(0, -5), path.join(componentDir, component)]
 }
 
 function toHtml(text: string) {
@@ -213,24 +66,35 @@ function toHtml(text: string) {
 	return ''
 }
 
-function logParseError(path: string, err: ParseError) {
-	if (err.name !== 'ParseError') {
-		throw err
+let currentRequest: express.Request
+function getHelpers(): {} {
+	const req = getCurrentRequest()
+	return {
+		config,
+		datetime,
+		fileHelpers,
+		i18n: req.__ ? req.__.bind(req) : null,
+		req,
+		signedInUser: req.user,
+		toHtml,
 	}
-	// console.error(`Error parsing ${path}:${err.loc.line}:${err.loc.column}`)
-	console.error()
-	console.error(err.frame)
-	console.error()
-	console.error(err.message)
-	console.error()
 }
+
+function getCurrentRequest() {
+	return currentRequest
+}
+
+/*function componentIdentity(component: string): [string, string] {
+	return [component.slice(0, -5), path.join(componentDir, component)]
+}
+
 function readComponentDir(dir: string, nestedDirName?: string) {
 	for (const file of fs.readdirSync(dir)) {
 		if (isDirectory(path.join(componentDir, file))) {
 			readComponentDir(path.join(componentDir, file), file)
 		}
 
-		if (file.endsWith('.html') && !file.includes('Aatest')) {
+		if (file.endsWith('.html')) {
 			// componentList.push(componentIdentity(file))
 
 			if (nestedDirName) {
@@ -246,83 +110,16 @@ function readComponentDir(dir: string, nestedDirName?: string) {
 	}
 }
 
-function resetCache() {
-	componentList = {}
-
+function readAll() {
 	readComponentDir(componentDir)
-	allComponentNames = Object.keys(componentList).join(', ')
-	components = {}
-	pages = {}
-	const reverseDeps: Record<string, Set<string>> = {}
-	const sources: Record<string, [string, string]> = {}
-	// Do a first pass in order to figure out the component dependencies.
-	for (const componentName of Object.keys(componentList)) {
-		const componentPath = componentList[componentName]
-		const source = fs.readFileSync(componentPath, {encoding: 'utf8'})
-		try {
-			const compiled = svelte.parse(source, {
-				css: false,
-				dev: false,
-				filename: componentPath,
-				format: 'cjs',
-				generate: 'ssr',
-				name: componentName,
-				shared: true,
-				onwarn: () => {
-					return
-				},
-			})
-			// TODO(tav): Handle cyclical dependencies.
-			for (const dep of getDependencies(compiled.html)) {
-				if (!reverseDeps[dep]) {
-					reverseDeps[dep] = new Set()
-				}
-				reverseDeps[dep].add(componentName)
-			}
-			sources[componentName] = [componentPath, source]
-		} catch (err) {
-			logParseError(componentPath, err)
-		}
-	}
-	// Do a second pass to figure out the right compilation order.
-	const seq: string[] = []
-	for (const componentName of Object.keys(componentList)) {
-		let idx = seq.length
-		const deps = reverseDeps[componentName]
-		if (deps) {
-			for (const dep of deps) {
-				const depIdx = seq.indexOf(dep)
-				if (depIdx === -1) {
-					continue
-				}
-				if (depIdx < idx) {
-					idx = depIdx
-				}
-			}
-		}
-		seq.splice(idx, 0, componentName)
-	}
-	// Do a final pass doing the actual Component compilation.
-	for (const componentName of seq) {
-		const [filepath, source] = sources[componentName]
-		components[componentName] = compile(filepath, source, componentName, true)
-	}
-}
+} */
 
 export function render(
 	page: string,
 	req: express.Request,
 	res: express.Response,
-	props?: any
-): string {
-	props = {...res.locals, ...props}
-	if (props && props.signedInUser) {
-		throw new Error('Attempt to override signedInUser in props')
-	}
-
-	let mod: Renderer | any | undefined = pages[page]
-	currentRequest = req
-
+	withData?: any
+) {
 	let pagePath = path.join(pageDir, page + '.html')
 	if (!isFile(pagePath)) {
 		pagePath = path.join(pageDir, page, 'index.html')
@@ -330,27 +127,17 @@ export function render(
 			throw new Error(`Could not find a matching .html file for ${page}`)
 		}
 	}
-	const source = fs.readFileSync(pagePath, {encoding: 'utf8'})
 
-	mod = compile(pagePath, source, getName(page))
-	if (!mod) {
-		throw new Error(`Could not compile renderer for ${page}`)
+	const component = require(pagePath)
+
+	currentRequest = req
+	logger.debug(`rendering page: ${page}`)
+
+	const data = {
+		...withData,
 	}
-	pages[page] = mod
-
-	// TODO(tav): Should use source maps to show any errors here in the context
-	// of the original HTML source as opposed to in the generated JavaScript
-	// code/module.
-
-	return mod.render(props).html
+	const store = new Store({
+		...getHelpers(),
+	})
+	res.send(component.render(data, {store}).html)
 }
-
-// TODO(tav): Populate the cache and clear it more intelligently, i.e. just the
-// ones that need to be regenerated as opposed to wiping the whole cache
-// whenever any template file changes.
-if (!config.PRODUCTION_ENV) {
-	fs.watch(componentDir, {}, resetCache)
-	fs.watch(pageDir, {recursive: true}, resetCache)
-}
-
-resetCache()
