+++ src/ui/component/Banner.html
@@ -7,5 +7,4 @@
     {/if}
 
     <slot></slot>
-
 </div>
