+++ src/ui/component/DisciteCourse.html
@@ -100,3 +100,4 @@
             {/if}
     </div>
 </div>
+
