+++ src/ui/page/test.html
@@ -1,10 +1,8 @@
-<PageTest>
-	<h1>hello test page!</h1>
-	{$config.ENV}
-</PageTest>
+<Page>
+</Page>
 
 <script>
-	import PageTest from '../component/PageTest.html';
+	import Page from '../component/Page.html';
 
 	export default {
 		data() {
@@ -13,7 +11,7 @@
 			}
 		},
 		components: {
-			PageTest
+			Page
 		}
 	};
 </script>
